fix: revert changes to sidebar
Mudanças na sidebar revertidas para o estado atual da main. As mudanças estão em um backup local para uma nova branch.
