+++ src/components/SideBar/sidebar.css
@@ -1,8 +1,10 @@
 #sideBarWrapper {
-    background-color: #F5EEDC;
+    background-color: #ffff;
     width: 95px;
-    height: 100vh;
+    height: 100%;
     border-right: 1px solid rgba(128, 128, 128, 0.5);
+    position: fixed;
+    z-index: 10;
 
     #sideBar {
         padding-top: 2rem;
@@ -18,6 +20,16 @@
         #proaeLogo {
             width: 55px;
             cursor: auto;
+        }
+
+        #option-wrapper {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+
+            p {
+                font-size: 0.8rem;
+            }
         }
 
         #iconWrapper {
